+++ packages/@codotype/generator-editor/src/components/OptionEditorLayout.vue
@@ -14,7 +14,7 @@
       </b-modal>
 
       <button class="btn btn-lg btn-primary btn-block" v-b-modal="'new-option'" @click="resetNewModel()">
-        <i class="fa fa-plus"></i>
+        <i class="fa fa-plus" />
         New Option
       </button>
     </template>