+++ packages/@codotype/ui/src/components/BlueprintEditor/components/attribute/AttributeDatatypeChild.vue
@@ -2,13 +2,13 @@
 <template>
   <div class="col-sm-6 my-2">
     <button class='btn btn-outline-dark btn-block text-left active' v-if="opt.value === val" @click="click(opt.value)">
-      <i :class="'fa-fw mr-2 ' + opt.icon"></i>
+      <i :class="'fa-fw mr-2 ' + opt.icon" />
       <strong>{{opt.label}}</strong>
       <small class='ml-2'>{{opt.desc}}</small>
     </button>
 
     <button class='btn btn-outline-dark btn-block text-left' v-else @click="click(opt.value)">
-      <i :class="'fa-fw mr-2 ' + opt.icon"></i>
+      <i :class="'fa-fw mr-2 ' + opt.icon" />
       <strong>{{opt.label}}</strong>
       <small class='ml-2'>{{opt.desc}}</small>
     </button>