+++ packages/@codotype/ui/CHANGELOG.md
@@ -7,6 +7,9 @@
 * Refactored `EditProjectModal` component, wraps `ProjectForm` component
 * Sized down `GenerateCodeButton`
 * Merged `ConfigurationEditor` component template and logic into `BlueprintEditor` component
+* Added `ProjectEditButton` component
+* Split `SchemaEditor` out of `BlueprintEditor` as a standalone component
+* Re-enabled `<TourButton />` component
 
 ---